Move the extension handling up a level and make them more generic
The rules for extension handling were imprecise (most only mentioned
ClientHello).  I've made that statement more generic.  It is also more
explicit, as it lists what messages contain relevant extensions.

There was considerable fuzziness here as it wasn't clear what to do if - for
example - oid_filters was present in the handshake but not the authenticator
request.  I've changed that so that it is clearer: if you have an authenticator
request, that is the only message that is relevant.