Focus on generic identities rather than certificates.
This changes a lot of text, since many places (perhaps rightfully so) assume the
contents of the Certificate message are a normal X509 certificate. Some uses
of EAs might just use the Certificate message as a vehicle for other identifying
information. Thus, rather than box ourselves into a corner, this change generalizes
the purpose of this Certificate to be that of an identity.